+++ wsgi.py
@@ -1,0 +1,3 @@
+from console import create_app
+
+app = create_app()